# docs/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details",
          "request_id"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "Body_upload_artifact_api_v1_artifacts_post": {
        "properties": {
          "artifact_type": {
            "default": "firmware",
            "title": "Artifact Type",
            "type": "string"
          },
          "ci_session_id": {
            "format": "uuid",
            "title": "Ci Session Id",
            "type": "string"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          }
        },
        "required": [
          "file",
          "ci_session_id"
        ],
        "title": "Body_upload_artifact_api_v1_artifacts_post",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "Phase 4 scoped API token.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Versioned local API for reserving and controlling lab benches.",
    "title": "Lab Platform Agent API",
    "version": "0.5.0-alpha"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/artifacts": {
      "post": {
        "operationId": "upload_artifact_api_v1_artifacts_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_artifact_api_v1_artifacts_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Upload Artifact Api V1 Artifacts Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Upload Artifact",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/api/v1/artifacts/{artifact_id}": {
      "get": {
        "operationId": "artifact_metadata_api_v1_artifacts__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Artifact Metadata Api V1 Artifacts  Artifact Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Artifact Metadata",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/api/v1/artifacts/{artifact_id}/content": {
      "get": {
        "operationId": "artifact_content_api_v1_artifacts__artifact_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Artifact file content."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Artifact Content",
        "tags": [
          "artifacts"
        ]
      }
    },
...
    "/api/v1/ci/sessions/{session_id}/artifacts": {
      "get": {
        "operationId": "session_artifacts_api_v1_ci_sessions__session_id__artifacts_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Session Artifacts Api V1 Ci Sessions  Session Id  Artifacts Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Session Artifacts",
        "tags": [
          "artifacts"
        ]
      }
    },
...
    "/api/v1/operations/{operation_id}/artifacts": {
      "get": {
        "operationId": "list_operation_artifacts_api_v1_operations__operation_id__artifacts_get",
        "parameters": [
          {
            "in": "path",
            "name": "operation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Operation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Operation Artifacts Api V1 Operations  Operation Id  Artifacts Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List Operation Artifacts"
      }
    },
...
  }
}